@@ -11,20 +11,24 @@
     "build": "rm -rf build/ && tsc",
     "start": "node build/index",
     "go": "npm run build && npm run start",
-    "lint": "prettier --ext=.ts,.js ./src"
+    "lint": "npx eslint ./src/**/*.ts"
   },
   "dependencies": {
     "cors": "^2.8.5",
-    "debug": "^4.3.1",
     "dotenv": "^10.0.0",
+    "envalid": "^7.1.0",
     "express": "^4.17.1",
     "express-jwt": "^6.0.0",
     "jsonwebtoken": "^8.5.1",
+    "jwt-decode": "^3.1.2",
     "mongoose": "^5.12.10",
+    "multer": "^1.4.2",
     "passport": "^0.4.1",
-    "passport-http": "^0.3.0",
+    "passport-jwt": "^4.0.0",
     "portfinder": "^1.0.28",
+    "redis": "^3.1.2",
     "socket.io": "^4.1.2",
+    "socketio-jwt-auth": "^0.2.1",
     "tslog": "^3.2.0"
   },
   "devDependencies": {
@@ -32,10 +36,19 @@
     "@types/express": "^4.17.11",
     "@types/express-jwt": "^6.0.1",
     "@types/jsonwebtoken": "^8.5.1",
+    "@types/mongoose": "^5.10.5",
+    "@types/multer": "^1.4.5",
     "@types/node": "^15.6.0",
     "@types/passport": "^1.0.6",
-    "typescript": "^4.2.4",
-    "prettier": "^2.3.0"
+    "@types/passport-jwt": "^3.0.5",
+    "@types/redis": "^2.8.29",
+    "@types/socketio-jwt-auth": "0.0.1",
+    "@typescript-eslint/eslint-plugin": "^4.26.0",
+    "@typescript-eslint/parser": "^4.26.0",
+    "eslint": "^7.28.0",
+    "eslint-config-prettier": "^8.3.0",
+    "prettier": "^2.3.1",
+    "typescript": "^4.3.2"
   },
   "private": true
 }
