@@ -1,35 +1,28 @@
-import mongoose = require('mongoose')
-import * as user from '../mongo/User'
-import { Logger } from 'tslog'
+import mongoose from 'mongoose'
+import { exit } from 'process'
+import logger from '../logger/'
+import { getUserByEmail, newUser, setAdmin } from '../mongo/userMethods'
 
-const logger = new Logger()
-
-export async function setupDB() {
-  // Connect to mongodb and launch the HTTP server trough Express
+export async function setupDB(): Promise<typeof mongoose> {
   try {
-    mongoose.set('useNewUrlParser', true)
-    mongoose.set('useFindAndModify', false)
-    mongoose.set('useCreateIndex', true)
-    mongoose.set('useUnifiedTopology', true)
-    const db = await mongoose.connect('mongodb://localhost:27017/connectfourdb')
+    const db = await mongoose.connect(process.env.MONGO_CONNECTION_STRING, {
+      useNewUrlParser: true,
+      useCreateIndex: true,
+      useUnifiedTopology: true,
+      useFindAndModify: false
+    })
     logger.info("Connected to MongoDB")
-    const doc = await user.getModel().findOne({ mail: "admin@postmessages.it" })
+    const doc = await getUserByEmail("admin@connectfour.it")
     if (!doc) {
       logger.info("Creating admin user")
-
-      var u = user.newUser({
-        username: "admin",
-        mail: "admin@connectfour.it"
-      })
-      u.setAdmin()
-      u.setModerator()
-      u.setPassword("admin")
-      u.save()
-    } else {
-      logger.info("Admin user already exists")
+      const admin = await newUser('admin', 'admin@connectfour.it', 'admin')
+      await setAdmin(admin.email)
     }
     return db
   } catch (err) {
-    logger.error(`Error Occurred during initialization: ${err}`)
+    logger.error(`Error occurred during DB initialization: ${err}`)
   }
+
+  logger.error(new Error('No db connection string found in process environment'))
+  exit(1)
 }
