@@ -1,15 +1,16 @@
-import { Player } from './Player'
-import { Logger } from 'tslog'
-
-const logger = new Logger()
-const rows = 6
-const columns = 7
+import logger from '../../logger'
+import MatchResults from '../../models/MatchResults'
+import { endMatch } from '../../mongo/matchMethods'
+import Player from '../Player'
+import MoveResult from './MoveResult'
 
 export class Match {
   public readonly player1: Player
   public readonly player2: Player
-  public readonly game_board = Array.from(Array(rows), _ => Array(columns).fill(0))
-  private heights = Array(columns).fill(0) //height of columns
+  private readonly rows = 6
+  private readonly columns = 7
+  public readonly game_board = Array.from(Array(this.rows), () => Array(this.columns).fill(0))
+  private heights = Array(this.columns).fill(0) //height of columns
   private p1Turn = true
 
   constructor(player1: Player, player2: Player) {
@@ -17,6 +18,11 @@ export class Match {
     this.player2 = player2
   }
 
+  /**
+   * 
+   * @param col indicates which column the user added a dot to
+   * @returns true if the move was a winning move, false otherwise
+   */
   private isWinner(col: number): boolean {
     if (this.heights[col] == 0) {
       logger.error("ERROR: Empty columns\n")
@@ -39,7 +45,10 @@ export class Match {
     for (let dx = -1; dx <= 1; dx++) {
       let nb = 0
       for (let dy = -1; dy <= 1; dy += 2) {
-        for (let x = this.heights[col] - 1 + dx * dy, y = col + dy; x >= 0 && x < rows && y >= 0 && y < columns && this.game_board[x][y] == player; nb++) {
+        for (let x = this.heights[col] - 1 + dx * dy, y = col + dy; x >= 0 &&
+          x < this.rows && y >= 0 &&
+          y < this.columns && this.game_board[x][y] == player; nb++) {
+
           x += dx * dy
           y += dy
         }
@@ -52,16 +61,49 @@ export class Match {
   }
 
   public colIsFull(col: number): boolean {
-    return (this.heights[col] == rows)
+    return (this.heights[col] == this.rows)
   }
 
-  public addDot(col: number, player: Player): boolean {
+  /**
+   * Received the move from a player, checks if the user is authorized to play and if it's that user's turn
+   * If the move is verified and accepted it calls isWinner to check if it's a winning move
+   * @param col Column number where the user inserted the dot
+   * @param playerId ID of the player who played the turn
+   * @returns a MoveResult object that specifies whether the move has been accepted and whether the move
+   * was a winning move through matchResult. If the property is set it contains the IDs of the winner and the loser
+   */
+  public addDot(col: number, playerId: string): MoveResult {
+    const res: MoveResult = {
+      accepted: false,
+      matchResult: undefined
+    }
+    let player: Player
+    if (this.player1.id === playerId) player = this.player1
+    else if (this.player2.id === playerId) player = this.player2
+    else return res
+
     if ((this.p1Turn && player === this.player1) || (!this.p1Turn && player === this.player2)) {
       this.game_board[this.heights[col]][col] = player.id
       this.heights[col]++
+      res.accepted = true
+
+      if (this.isWinner(col)) {
+        let loser: Player
+        if (player === this.player1) loser = this.player2
+        else loser = this.player1
+        res.matchResult = {
+          winner: player.username,
+          loser: loser.username
+        }
+
+        this.endGame(res.matchResult)
+      }
       this.p1Turn = !this.p1Turn
-      return this.isWinner(col)
     }
-    return false
+    return res
+  }
+
+  private endGame(res: MatchResults) {
+    endMatch(res)
   }
 }
