@@ -1,36 +1,42 @@
-import { Server as IOServer } from 'socket.io'
-import { Logger } from 'tslog'
+import { Socket } from 'socket.io'
+import logger from '../../logger/'
+import { createIOServer } from '../../setup/setupIOServer'
 import freePortFinder from '../../utils/freePortFinder'
-import { Player } from './Player'
+import UnmatchedPlayer from '../matchmaking/UnmatchedPlayer'
+import Player from '../Player'
+import { matchCallback } from '../socket/matchSocket'
 import { Match } from './Match'
 
-const logger = new Logger()
-
-export async function gameStart(p1: Player, p2: Player) {
+/**
+ * Creates a new Socket.io server and starts the game
+ * @param p1 player 1
+ * @param p2 player 2
+ */
+export async function gameStart(p1: UnmatchedPlayer, p2: UnmatchedPlayer): Promise<void> {
   const port = await freePortFinder()
 
   if (port) {
-    const match = new Match(p1, p2)
-    const io = new IOServer(port, { cors: { origin: "*" } })
+    const io = createIOServer(port)
 
-    p1.ws.emit('matched', port)
-    p2.ws.emit('matched', port)
+      ; (p1.ws as Socket).emit('matched', port)
+      ; (p2.ws as Socket).emit('matched', port)
 
-    io.on('connection', (socket) => {
-      logger.info(`Started a new match between ${p1.id} and ${p2.id}`)
-      p1.ws = socket
-      p2.ws = socket
+    let player1: Player, player2: Player
+    if (Math.random()) {
+      player1 = p1.player
+      player2 = p2.player
+    } else {
+      player1 = p2.player
+      player2 = p1.player
+    }
 
-      socket.on('move', (message) => {
-        if (match.addDot(message.col, message.player)) {
-          socket.broadcast.emit(`Player ${message.player} has won the match!`)
-          io.disconnectSockets()
-        }
-      })
-    })
+    const match = new Match(player1, player2)
+      ; (p1.ws as Socket).broadcast.emit('playing', p1.player.username, port)
+      ; (p2.ws as Socket).broadcast.emit('playing', p2.player.username, port)
+
+    logger.info(`Started a new match between ${player1.username} and ${player2.username}`)
+    io.on('connection', socket => { matchCallback(match, p1, p2, io, socket) })
   } else {
     logger.error("Couldn't find a free port")
-    p1.ws.disconnect()
-    p2.ws.disconnect()
   }
 }
