@@ -1,5 +1,8 @@
-export interface UnmatchedPlayer {
-  id: number
-  mmr: number
+import { Socket } from "socket.io"
+import Player from "../Player"
+
+export default interface UnmatchedPlayer {
+  player: Player
   timeJoined: number
+  ws: Socket | string
 }
