@@ -1,7 +1,8 @@
 import express from 'express'
-import controller from '../controllers/health'
+import { healthCheck } from '../controllers/health'
+
 const router = express.Router()
 
-router.get('/', controller.healthCheck)
+router.get('/', healthCheck)
 
-export = router
+export default router
