@@ -1,7 +1,10 @@
 import express from 'express'
-import controller from '../controllers/login'
+import { login } from '../controllers/login'
+import { newPassword } from '../controllers/userPassword'
+
 const router = express.Router()
 
-router.get("/login", controller.login)
+router.post('/', express.urlencoded({ extended: true }), express.json(), login)
+router.post('/first', express.urlencoded({ extended: true }), express.json(), newPassword)
 
-export = router
+export default router
