@@ -1,9 +1,7 @@
-import {Logger} from 'tslog'
-import {Request, Response, NextFunction} from 'express'
+import { NextFunction, Request, Response } from 'express'
+import logger from '../../logger'
 
-const logger = new Logger()
-
-export function requestLogger(req: Request, res: Response, next: NextFunction) {
+export function requestLogger(req: Request, _res: Response, next: NextFunction): void {
   logger.info(`[METHOD]: ${req.method} - [ENDPOINT]: ${req.path}`)
   next()
 }
